Introduce DLCS.AWS project and move S3 classes
