+++ API/Features/Image/Requests/IngestImageFromFile.cs
@@ -5,8 +5,9 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using API.Settings;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Core;
-using DLCS.Model.Storage;
 using DLCS.Web.Requests;
 using MediatR;
 using Microsoft.Extensions.Logging;
