+++ API/Startup.cs
@@ -4,13 +4,12 @@ using API.Auth;
 using API.Client;
 using API.Infrastructure;
 using API.Settings;
+using DLCS.AWS.S3;
 using DLCS.Core.Encryption;
 using DLCS.Model.Customers;
-using DLCS.Model.Storage;
 using DLCS.Repository;
 using DLCS.Repository.Caching;
 using DLCS.Repository.Customers;
-using DLCS.Repository.Storage.S3;
 using DLCS.Web.Configuration;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
