+++ DLCS.Repository.Tests/Storage/S3/BucketReaderTests.cs
@@ -7,9 +7,9 @@ using System.Threading.Tasks;
 using Amazon.Runtime;
 using Amazon.S3;
 using Amazon.S3.Model;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Core.Exceptions;
-using DLCS.Model.Storage;
-using DLCS.Repository.Storage.S3;
 using FakeItEasy;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
