+++ DLCS.Repository.Tests/Strategy/S3AmbientOriginStrategyTests.cs
@@ -2,9 +2,10 @@
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Core.Types;
 using DLCS.Model.Customers;
-using DLCS.Model.Storage;
 using DLCS.Repository.Strategy;
 using FakeItEasy;
 using FluentAssertions;
