+++ DLCS.Repository/Assets/ThumbReorganiser.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Core.Collections;
 using DLCS.Core.Threading;
 using DLCS.Model.Assets;
-using DLCS.Model.Storage;
 using DLCS.Repository.Storage;
 using IIIF;
 using Microsoft.Extensions.Logging;
