+++ DLCS.Repository/Auth/DapperCredentialsRepository.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.IO;
 using System.Threading.Tasks;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Core.Guard;
 using DLCS.Model.Auth;
 using DLCS.Model.Customers;
-using DLCS.Model.Storage;
 using DLCS.Repository.Caching;
 using LazyCache;
 using Microsoft.Extensions.Caching.Memory;
