+++ Orchestrator.Tests/Infrastructure/NamedQueries/PDF/FireballPdfCreatorTests.cs
@@ -4,10 +4,11 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Model.Assets;
 using DLCS.Model.Assets.NamedQueries;
 using DLCS.Model.PathElements;
-using DLCS.Model.Storage;
 using FakeItEasy;
 using FizzWare.NBuilder;
 using FluentAssertions;
