+++ Orchestrator/Infrastructure/NamedQueries/PDF/FireballPdfCreator.cs
@@ -6,10 +6,10 @@ using System.Net.Http;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using DLCS.AWS.S3;
 using DLCS.Core.Strings;
 using DLCS.Model.Assets;
 using DLCS.Model.Assets.NamedQueries;
-using DLCS.Model.Storage;
 using DLCS.Repository.Storage;
 using DLCS.Web.Response;
 using Microsoft.Extensions.Logging;
