+++ Orchestrator/Infrastructure/NamedQueries/Zip/ImageThumbZipCreator.cs
@@ -5,9 +5,10 @@ using System.IO.Compression;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Model.Assets;
 using DLCS.Model.Assets.NamedQueries;
-using DLCS.Model.Storage;
 using DLCS.Repository.Storage;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
