+++ Portal/Features/Images/Requests/IngestSingleImage.cs
@@ -3,9 +3,10 @@ using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 using API.Client;
+using DLCS.AWS.S3;
+using DLCS.AWS.S3.Models;
 using DLCS.Core.Settings;
 using DLCS.HydraModel;
-using DLCS.Model.Storage;
 using DLCS.Repository.Spaces;
 using MediatR;
 using Microsoft.Extensions.Logging;
