+++ Portal/Startup.cs
@@ -4,13 +4,12 @@ using System.Net.Http.Headers;
 using System.Security.Claims;
 using Amazon.S3;
 using API.Client;
+using DLCS.AWS.S3;
 using DLCS.Core.Encryption;
 using DLCS.Core.Settings;
 using DLCS.Mediatr.Behaviours;
-using DLCS.Model.Storage;
 using DLCS.Repository;
 using DLCS.Repository.Spaces;
-using DLCS.Repository.Storage.S3;
 using MediatR;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authorization;
