+++ protagonist.sln
@@ -49,6 +49,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hydra", "Hydra\Hydra.csproj
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DLCS.Mock", "DLCS.Mock\DLCS.Mock.csproj", "{963A0ED4-2836-42F0-9254-E405626EA7CE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DLCS.AWS", "DLCS.AWS\DLCS.AWS.csproj", "{5AE90CE9-26E2-4C4B-85F9-78BC4F406667}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -139,6 +141,10 @@ Global
 		{963A0ED4-2836-42F0-9254-E405626EA7CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{963A0ED4-2836-42F0-9254-E405626EA7CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{963A0ED4-2836-42F0-9254-E405626EA7CE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5AE90CE9-26E2-4C4B-85F9-78BC4F406667}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5AE90CE9-26E2-4C4B-85F9-78BC4F406667}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5AE90CE9-26E2-4C4B-85F9-78BC4F406667}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5AE90CE9-26E2-4C4B-85F9-78BC4F406667}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
